requirements.txt
# Core dependencies
requests>=2.31.0

# Optional: Browser automation for Mercari (if not using Shops API)
playwright>=1.40.0

# Optional: Image processing
Pillow>=10.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
